requirements.txt
django~=4.2.29
djangorestframework~=3.16.1
django-model-utils~=5.0.0
